Move the call to the primary dialog box. Change the values of the data received from the API to Response<>
