+++ data/src/main/java/com/example/data/room/ArtsService.kt
@@ -2,6 +2,7 @@ package com.example.data.room
 
 
 import dataclasses.arts.Artworks
+import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Query
 
@@ -11,5 +12,5 @@ interface ArtsService {
     @GET("artworks")
     suspend fun getAllArts(
         @Query("page") currentPage: Int?
-    ): Artworks
+    ): Response<Artworks>
 }
